re-organize code to resolve #33
oops
